AppTime, 일별, 주별 ,월별 CRUD에 모두 jwt 토큰 적용. Goal, SelecteApp DB 추가 및 jwt 토큰 적용

+++ src/main/java/com/example/ATM_Backend/appTime/controller/MonthlyUsageController.java
@@ -2,6 +2,10 @@ package com.example.ATM_Backend.appTime.controller;
 
 import com.example.ATM_Backend.appTime.model.MonthlyUsage;
 import com.example.ATM_Backend.appTime.service.MonthlyUsageService;
+import com.example.ATM_Backend.jwt.JwtTokenProvider;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.ExampleObject;
+import io.swagger.v3.oas.annotations.media.Schema;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +18,9 @@ import java.util.List;
 public class MonthlyUsageController {
 
     private final MonthlyUsageService monthlyUsageService;
-
+    @Autowired
+    private JwtTokenProvider jwtTokenProvider;
+    
     @Autowired
     public MonthlyUsageController(MonthlyUsageService monthlyUsageService) {
         this.monthlyUsageService = monthlyUsageService;
@@ -31,21 +37,64 @@ public class MonthlyUsageController {
     }
 
     @GetMapping("/get/{userName}")
-    public ResponseEntity<List<MonthlyUsage>> getMonthlyUsageByUserName(@PathVariable("userName") String userName) {
+    public ResponseEntity<List<MonthlyUsage>> getMonthlyUsageByUserName(
+            @Parameter(description = "Authorization Token", required = true,
+                    examples = @ExampleObject(name = "Authorization 예시", value = "사용자 jwt 토큰"),
+                    schema = @Schema(type = "string"))
+            @RequestHeader(value = "Authorization") String token,
+            @PathVariable("userName") String userName) {
+
         try {
-            List<MonthlyUsage> monthlyUsage = monthlyUsageService.getMonthlyUsageByUserName(userName);
-            if (monthlyUsage.isEmpty()) {
+            if (token.startsWith("Bearer ")) {
+                token = token.substring(7);
+            } else {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+            }
+
+            if (!jwtTokenProvider.validateToken(token)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+            }
+
+            String userPK = jwtTokenProvider.getUserPK(token);
+            if (!userName.equals(userPK)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+            }
+
+            List<MonthlyUsage> monthlyUsages = monthlyUsageService.getMonthlyUsageByUserName(userName);
+            if (monthlyUsages.isEmpty()) {
                 return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
             }
-            return ResponseEntity.ok(monthlyUsage);
+            return ResponseEntity.ok(monthlyUsages);
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
         }
     }
 
+
     @DeleteMapping("/delete/{userName}")
-    public ResponseEntity<String> deleteMonthlyUsageById(@PathVariable("userName") String userName) {
+    public ResponseEntity<String> deleteMonthlyUsageByUserName(
+            @Parameter(description = "Authorization Token", required = true,
+                    examples = @ExampleObject(name = "Authorization 예시", value = "사용자 jwt 토큰"),
+                    schema = @Schema(type = "string"))
+            @RequestHeader(value = "Authorization") String token,
+            @PathVariable("userName") String userName) {
+
         try {
+            if (token.startsWith("Bearer ")) {
+                token = token.substring(7);
+            } else {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("유효하지 않은 토큰 형식입니다.");
+            }
+
+            if (!jwtTokenProvider.validateToken(token)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("유효하지 않은 토큰입니다.");
+            }
+
+            String userPK = jwtTokenProvider.getUserPK(token);
+            if (!userName.equals(userPK)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("토큰이 사용자의 userName과 일치하지 않습니다.");
+            }
+
             boolean isDeleted = monthlyUsageService.deleteMonthlyUsageByUserName(userName);
             if (isDeleted) {
                 return ResponseEntity.ok("MonthlyUsage deleted successfully");
